+++ backend/Polyglot.BusinessLogic/Infrastructure.cs
@@ -20,7 +20,7 @@ namespace Polyglot.BusinessLogic
             services.AddTransient<IComplexStringService, ComplexStringService>();
             services.AddTransient<IUserService, UserService>();
             services.AddTransient<ITeamService, TeamsService>();
-
+            services.AddTransient<IRightService, RightService>();
             services.AddTransient<IGlossaryService, GlossaryService>();
 
             services.AddTransient<IRatingService, RatingService>();
