+++ backend/Polyglot.BusinessLogic/Services/TeamsService.cs
@@ -236,48 +236,6 @@ namespace Polyglot.BusinessLogic.Services
             return 0.0d;
         }
 
-        public Task<IEnumerable<RightDTO>> GetTranslatorRightsAsync(int translatorId)
-        {
-// TODO IMPLEMENT
-            throw new System.NotImplementedException();
-        }
-
-        public async Task<TranslatorDTO> SetTranslatorRight(int userId, int teamId, RightDefinition definition)
-        {
-            var team = await uow.GetRepository<Team>().GetAsync(teamId);
-            var translator = team.TeamTranslators.FirstOrDefault(t => t.UserProfile.Id == userId);
-
-            var right = (await uow.GetRepository<Right>().GetAllAsync())
-                    .FirstOrDefault(r => r.Definition == definition);
-            translator.TranslatorRights.Add(new TranslatorRight()
-            {
-                RightId = right.Id,
-                TeamTranslatorId = translator.Id
-            });
-
-            var newTranslator = uow.GetRepository<TeamTranslator>().Update(translator);
-            await uow.SaveAsync();
-
-            return newTranslator != null ? mapper.Map<TranslatorDTO>(newTranslator) : null;
-        }
-
-        public async Task<TranslatorDTO> RemoveTranslatorRight(int userId, int teamId, RightDefinition definition)
-        {
-            var team = await uow.GetRepository<Team>().GetAsync(teamId);
-            var translator = team.TeamTranslators.FirstOrDefault(t => t.UserProfile.Id == userId);
-
-            var right = (await uow.GetRepository<Right>().GetAllAsync())
-                    .FirstOrDefault(r => r.Definition == definition);
-            var translatorRight = translator.TranslatorRights
-                .FirstOrDefault(tr => tr.RightId == right.Id &&  tr.TeamTranslatorId == translator.Id);
-
-            translator.TranslatorRights.Remove(translatorRight);
-
-            var newTranslator = uow.GetRepository<TeamTranslator>().Update(translator);
-            await uow.SaveAsync();
-
-            return newTranslator != null ? mapper.Map<TranslatorDTO>(newTranslator) : null;
-        }
 
         #endregion Translators
     }
